--- src/Application/Common/MappingProfiles/RequestProfiles.cs
@@ -1,7 +1,0 @@
-﻿namespace Application.Common.MappingProfiles;
-
-/// <summary>
-/// The <see langword="static"/> request profiles class.
-/// </summary>
-internal static partial class RequestProfiles
-{ }
