--- src/Application/Common/MappingProfiles/ResponseProfiles.cs
@@ -1,7 +1,0 @@
-﻿namespace Application.Common.MappingProfiles;
-
-/// <summary>
-/// The <see langword="static"/> response profiles class.
-/// </summary>
-internal static partial class ResponseProfiles
-{ }
