+++ src/Application/Contracts/Requests/Documents/DocumentUpdateRequest.cs
@@ -1,16 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
+
+using Application.Contracts.Requests.Documents.Base;
 
 namespace Application.Contracts.Requests.Documents;
 
 /// <summary>
-/// The request to update a document.
+/// The request for updating a document.
 /// </summary>
-public sealed class DocumentUpdateRequest
+public sealed class DocumentUpdateRequest : DocumentBaseRequest
 {
 	/// <summary>
 	/// The unique identifier of the document.
